WMS layer ordering bug fix from Tore.  Updated the rest of the projections to make integer pixel offset a double offset.  Removed LLXYView.
git-svn-id: svn+ssh://src.dsl.bbn.com/svn/openmap/trunk@1966 f68c23cc-7f87-4fbc-acfa-d9760b90acc1

+++ src/openmap/com/bbn/openmap/proj/Gnomonic.java
@@ -39,7 +39,7 @@ public class Gnomonic extends Azimuth {
      */
     public final static transient String GnomonicName = "Gnomonic";
 
-    protected int hy, wx;
+    protected double hy, wx;
 
     // almost constant projection parameters
     protected double cosCtrLat;
